requirements.txt
numpy==1.18.2
pillow==7.0.0
matplotlib==3.1.3
tensorflow==2.1.0
tensorflow_hub==0.7.0
imageio==2.8.0
jupyter==1.0.0